pyproject.toml
[project]
name = "phantomdb"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "click",
    "pandas",
    "sqlalchemy",
    "pydicom",
    "bs4",
    "atlassian-python-api",
    "requests",
    "attrs",
    "lxml",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
phantomdb = "phantomdb.app:main"
